quant-trading-system/requirements.txt
# 量化交易系统依赖

# FastAPI框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
aiomysql==0.2.0
redis==5.0.1
clickhouse-driver==0.2.7
clickhouse-connect==0.7.3

# 数据科学
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
ta-lib==0.4.28

# 金融数据源
akshare>=1.10.0
tushare>=1.2.0

# 缓存与限流
aiocache>=0.12.0
slowapi>=0.1.0

# 消息队列
kafka-python==2.0.2
pika==1.3.2
celery==5.3.4

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0
pandas-ta==0.3.14b0

# 网络请求
httpx==0.25.1
aiohttp==3.9.1
websockets==12.0

# 日期时间处理
python-dateutil==2.8.2
pytz==2023.3.post1

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# 监控与日志
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0